Move `ContentModeratorController` to the content moderator controllers folder
